Add argument to checkLayout() so it doesn't call done() always
This is useful for some tests that call checkLayout() several times,
so we don't need to use the old Blink check-layout.js and
we can move to the new check-layout-th.js.

Modified 3 tests to start using check-layout-th.js thanks
to this change.

Updated the copies of check-layout-th.js in WPT test suites.

Change-Id: Icfe696a1e43bd1c69e43f05eb0682cf599af52f1